Move configmap functions into overcloudipset package
