+++ controllers/overcloudipset_controller.go
@@ -17,9 +17,7 @@ limitations under the License.
 package controllers
 
 import (
-	"bytes"
 	"context"
-	"fmt"
 	"net"
 
 	"github.com/go-logr/logr"
@@ -32,6 +30,7 @@ import (
 
 	ospdirectorv1beta1 "github.com/openstack-k8s-operators/osp-director-operator/api/v1beta1"
 	common "github.com/openstack-k8s-operators/osp-director-operator/pkg/common"
+	overcloudipset "github.com/openstack-k8s-operators/osp-director-operator/pkg/overcloudipset"
 )
 
 // OvercloudIPSetReconciler reconciles a OvercloudIPSet object
@@ -148,41 +147,11 @@ func (r *OvercloudIPSetReconciler) Reconcile(req ctrl.Request) (ctrl.Result, err
 		return ctrl.Result{}, err
 	}
 
-	// network isolation w/ predictable IPs:
-	//	   https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/features/network_isolation.html
-	//     https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/provisioning/node_placement.html#predictable-ips
-	roleIPSets := map[string]map[string][]string{}
-	// DeployedServerPortMap:
-	//     https://docs.openstack.org/project-deploy-guide/tripleo-docs/latest/features/deployed_server.html
-	ctlPlaneIps := map[string]string{}
-
-	for _, ipset := range overcloudIPList.Items {
-		for netName, addr := range ipset.Status.IPAddresses {
-			if netName == "ctlplane" {
-				ctlPlaneIps[ipset.Name] = addr
-			} else {
-				if roleIPSet, exists := roleIPSets[ipset.Spec.Role]; exists {
-					//roleIPSets[ipset.Spec.Role] = map[string][]string{netName: append(roleIPSet[netName], addr)}
-					roleIPSet[netName] = append(roleIPSet[netName], addr)
-				} else {
-					roleIPSets[ipset.Spec.Role] = map[string][]string{netName: {addr}}
-				}
-			}
-		}
-	}
-
-	r.Log.Info("RoleIPList:")
-	r.Log.Info(fmt.Sprintf("%v", roleIPSets))
-	r.Log.Info(fmt.Sprintf("Ctlplane CIDR: %s", ctlplaneCidr))
-
 	// write it all to a configmap
 	envVars := make(map[string]common.EnvSetter)
 	cmLabels := common.GetLabels(instance.Name, "osp-overcloudipset")
 
-	templateParameters := make(map[string]string)
-	templateParameters["PredictableIps"] = roleIPSetToString(roleIPSets)
-	templateParameters["DeployedServerPortMap"] = deployedServerPortMap(ctlPlaneIps, ctlplaneCidr)
-	templateParameters["CtlplaneCidr"] = ctlplaneCidr
+	templateParameters := overcloudipset.CreateConfigMapParams(*overcloudIPList, ctlplaneCidr)
 
 	cm := []common.Template{
 		{
@@ -210,32 +179,4 @@ func (r *OvercloudIPSetReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&ospdirectorv1beta1.OvercloudIPSet{}).
 		Complete(r)
-}
-
-// roleIPSetToString -
-func roleIPSetToString(r map[string]map[string][]string) string {
-	var b bytes.Buffer
-	for role, roleIps := range r {
-		b.WriteString(fmt.Sprintf("  %sIPs:\n", role))
-		for netname, ips := range roleIps {
-			b.WriteString(fmt.Sprintf("    %s:\n", netname))
-			for _, ip := range ips {
-				b.WriteString(fmt.Sprintf("    - %s\n", ip))
-			}
-		}
-	}
-	return b.String()
-}
-
-// deployedServerPortMap -
-func deployedServerPortMap(r map[string]string, c string) string {
-	var b bytes.Buffer
-	for hostname, ipaddr := range r {
-		b.WriteString(fmt.Sprintf("    %s-ctlplane:\n", hostname))
-		b.WriteString(fmt.Sprintf("      fixed_ips:\n"))
-		b.WriteString(fmt.Sprintf("        - ip_address: %s\n", ipaddr))
-		b.WriteString(fmt.Sprintf("      subnets:\n"))
-		b.WriteString(fmt.Sprintf("        - cidr: %s\n", c))
-	}
-	return b.String()
 }
